Adding insert_mode switch for fact and dimension table inserts (delete+insert or append). Modifying SqlQueries class with right queries. Adding insert_mode parameter to etl_dag, LoadFactOperator, LoadDimensionOperator.

+++ plugins/operators/load_dimension.py
@@ -11,6 +11,15 @@ class LoadDimensionOperator(BaseOperator):
         INSERT INTO {} ({})
         {};
     """
+    # sql_template_append = """
+    #     INSERT INTO {} ({})
+    #     {}
+    #     AND NOT EXISTS(
+    #         SELECT {}
+    #         FROM {}
+    #         WHERE {}.{} = {}.{}
+    #     );
+    # """
     @apply_defaults
     def __init__(self,
                  # Define operators params (with defaults)
@@ -18,6 +27,7 @@ class LoadDimensionOperator(BaseOperator):
                  target_table="",
                  target_columns="",
                  query="",
+                 insert_mode="append",
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
@@ -26,25 +36,45 @@ class LoadDimensionOperator(BaseOperator):
         self.target_table = target_table
         self.target_columns = target_columns
         self.query = query
+        self.insert_mode = insert_mode
 
     def execute(self, context):
         # Set AWS Redshift connections
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        self.log.info("Clearing data from Redshift target table {}..."\
-                        .format(self.target_table))
-        redshift.run("DELETE FROM {}".format(self.target_table))
+        if self.insert_mode == "truncate_insert":
+            self.log.info("Insert_mode = truncate_insert. Clearing data from Redshift target table: {} ..."\
+                            .format(self.target_table))
+            redshift.run("DELETE FROM {}".format(self.target_table))
+        elif self.insert_mode == "append":
+            self.log.info("Insert_mode = append. Inserting new data on top of old one in Redshift target table: {} ..."\
+                            .format(self.target_table))
+        else:
+            self.log.info("Insert_mode not defined => using append (default value). Inserting new data on top of old one in Redshift target table: {} ..."\
+                            .format(self.target_table))
 
         self.log.info("Preparing SQL query for {} table".format(self.target_table))
         query_name = ""
         if self.query == "user_table_insert":
-            query_name = SqlQueries.user_table_insert
+            if self.insert_mode == "append":
+                query_name = SqlQueries.user_table_insert_append
+            else:
+                query_name = SqlQueries.user_table_insert_delete
         elif self.query == "song_table_insert":
-            query_name = SqlQueries.song_table_insert
+            if self.insert_mode == "append":
+                query_name = SqlQueries.song_table_insert_append
+            else:
+                query_name = SqlQueries.song_table_insert_delete
         elif self.query == "artist_table_insert":
-            query_name = SqlQueries.artist_table_insert
+            if self.insert_mode == "append":
+                query_name = SqlQueries.artist_table_insert_append
+            else:
+                query_name = SqlQueries.artist_table_insert_delete
         elif self.query == "time_table_insert":
-            query_name = SqlQueries.time_table_insert
+            if self.insert_mode == "append":
+                query_name = SqlQueries.time_table_insert_append
+            else:
+                query_name = SqlQueries.time_table_insert_delete
         else:
             query_name = ""
             self.log.info("Invalid value in query parameter.")
@@ -52,7 +82,7 @@ class LoadDimensionOperator(BaseOperator):
         formatted_sql = LoadDimensionOperator.sql_template.format(
             self.target_table,
             self.target_columns,
-            query_name
+            query_name.format(self.target_table, self.target_table)
         )
 
         # Executing Load operation
